Consolidated updating and rolling back to a single method.
Initially, rollback() was one method and update() was another method,
however now they've been combined to a single method update($version=-1)
to update the database to any version at any time. If $version is -1,
it is assumed to update to the HEAD version. If $version is greater than
the current most installed version in the database, an update
occurs with the migration scripts between $currentVersion and $version.
If $version is less than the current most intalled version in the database,
a rollback occurs, returing the database to $version.