Stores the setUp and tearDown queries in the database.
Rather than building a bunch of setUp and tearDown files on the disk,
the query is stored in the database. This way, if you change to a
version of the repository that is behind the current version in the
database and you don't have the migration scripts, you can still
rollback to the most current version.